make admin

+++ database/migrations/2021_07_12_041123_create_courses_table.php
@@ -26,7 +26,7 @@ class CreateCoursesTable extends Migration
             $table->integer("instructor_id");
             $table->string("instructor_name");
             $table->integer("number_of_lessons")->default(0);
-            $table->float("total_enrollment")->default(0.00);
+            $table->integer("total_enrollment")->default(0);
             $table->float("total_earning_course")->default(0.00);
             $table->timestamps();
         });
